feat: 初步完成raft对外接入重构，把与外部的数据交互汇总在raftdata::RaftDataHandler中
