+++ integration_tests/rust-client-test/tests/config/config_cluster_tests.rs
@@ -580,7 +580,7 @@ async fn test_config_batch_add_query(
         let key = ConfigKey::new(data_id, &group, "");
         add_client.set_config(&key, value).await.unwrap();
     }
-    
+
     tokio::time::sleep(Duration::from_millis(100)).await;
 
     // Query all configs
